Add files via upload
Corrected the position of the date and labels in properties. Now it is aligned with the other values

A small correction in the embedded notes (the bottom padding was too short).

Tables show vertical separator lines inside.

New option to separately choose font size inside code blocks in edit mode and preview mode.